src/index.ts
import * as builderts from './builder'
import * as configts from './config'
import * as factoryts from './factory'
import * as optionsts from './options'
import * as serviceservicets from './service/service'
import * as servicevolumets from './service/volume'
import * as servicewithts from './service/with'
import * as versionts from './version'
export { builderts as builder }
export { configts as config }
export { factoryts as factory }
export { optionsts as options }
export const service = {
  service: serviceservicets,
  volume: servicevolumets,
  with: servicewithts
}
export { versionts as version }
